Move common to internal
