+++ experimental/auth-logic/wrappers/BUILD
@@ -46,7 +46,7 @@ go_library(
     importpath = "github.com/project-oak/transparent-release/experimental/auth-logic/wrappers",
     deps = [
         "//pkg/amber",
-        "//common",
+        "//internal/common",
         "//verify",
         "@com_github_in_toto_in_toto_golang//in_toto:go_default_library",
         "@com_github_cyberphone_json_canonicalization//go/src/webpki.org/jsoncanonicalizer",
